Conflicts:
	octoprint/printer.py
	octoprint/server.py
	octoprint/util/comm.py

+++ octoprint/server.py
@@ -126,7 +126,8 @@ def index():
 		enableTimelapse=(settings().get(["webcam", "snapshot"]) is not None and settings().get(["webcam", "ffmpeg"]) is not None),
 		enableGCodeVisualizer=settings().get(["feature", "gCodeVisualizer"]),
 		enableSystemMenu=settings().get(["system"]) is not None and settings().get(["system", "actions"]) is not None and len(settings().get(["system", "actions"])) > 0,
-		enableAccessControl=userManager is not None
+		enableAccessControl=userManager is not None,
+		enableSdSupport=settings().get(["feature", "sdSupport"])
 	)
 
 #~~ Printer control
@@ -261,11 +262,40 @@ def getCustomControls():
 	customControls = settings().get(["controls"])
 	return jsonify(controls=customControls)
 
+@app.route(BASEURL + "control/sd", methods=["POST"])
+@login_required
+def sdCommand():
+	if not settings().getBoolean(["feature", "sdSupport"]) or not printer.isOperational() or printer.isPrinting():
+		return jsonify(SUCCESS)
+
+	if "command" in request.values.keys():
+		command = request.values["command"]
+		if command == "init":
+			printer.initSdCard()
+		elif command == "refresh":
+			printer.refreshSdFiles()
+		elif command == "release":
+			printer.releaseSdCard()
+
+	return jsonify(SUCCESS)
+
 #~~ GCODE file handling
 
 @app.route(BASEURL + "gcodefiles", methods=["GET"])
 def readGcodeFiles():
-	return jsonify(files=gcodeManager.getAllFileData())
+	files = gcodeManager.getAllFileData()
+
+	sdFileList = printer.getSdFiles()
+	if sdFileList is not None:
+		for sdFile in sdFileList:
+			files.append({
+				"name": sdFile,
+				"size": "n/a",
+				"bytes": 0,
+				"date": "n/a",
+				"origin": "sd"
+			})
+	return jsonify(files=files)
 
 @app.route(BASEURL + "gcodefiles/<path:filename>", methods=["GET"])
 def readGcodeFile(filename):
@@ -278,6 +308,8 @@ def uploadGcodeFile():
 	if "gcode_file" in request.files.keys():
 		file = request.files["gcode_file"]
 		filename = gcodeManager.addFile(file)
+		if filename and "target" in request.values.keys() and request.values["target"] == "sd":
+			printer.addSdFile(filename, gcodeManager.getAbsolutePath(filename))
 
 		global eventManager
 		eventManager.fire("Upload", filename)
@@ -290,12 +322,17 @@ def loadGcodeFile():
 		printAfterLoading = False
 		if "print" in request.values.keys() and request.values["print"] in valid_boolean_trues:
 			printAfterLoading = True
-		filename = gcodeManager.getAbsolutePath(request.values["filename"])
-		if filename is not None:
-			printer.loadGcode(filename, printAfterLoading)
 
-			global eventManager
-			eventManager.fire("LoadStart", filename)
+		if "target" in request.values.keys() and request.values["target"] == "sd":
+			filename = request.values["filename"]
+			printer.selectSdFile(filename, printAfterLoading)
+		else:
+			filename = gcodeManager.getAbsolutePath(request.values["filename"])
+			if filename is not None:
+				printer.loadGcode(filename, printAfterLoading)
+
+                global eventManager
+                eventManager.fire("LoadStart", filename)
 	return jsonify(SUCCESS)
 
 @app.route(BASEURL + "gcodefiles/delete", methods=["POST"])
@@ -303,8 +340,16 @@ def loadGcodeFile():
 def deleteGcodeFile():
 	if "filename" in request.values.keys():
 		filename = request.values["filename"]
-		gcodeManager.removeFile(filename)
+		if "target" in request.values.keys() and request.values["target"] == "sd":
+			printer.deleteSdFile(filename)
+		else:
+			gcodeManager.removeFile(filename)
 	return readGcodeFiles()
+
+@app.route(BASEURL + "gcodefiles/refresh", methods=["POST"])
+def refreshFiles():
+	printer.updateSdFiles()
+	return jsonify(SUCCESS)
 
 #~~ timelapse handling
 
@@ -394,7 +439,10 @@ def getSettings():
 		},
 		"feature": {
 			"gcodeViewer": s.getBoolean(["feature", "gCodeVisualizer"]),
-			"waitForStart": s.getBoolean(["feature", "waitForStartOnConnect"])
+			"waitForStart": s.getBoolean(["feature", "waitForStartOnConnect"]),
+			"alwaysSendChecksum": s.getBoolean(["feature", "alwaysSendChecksum"]),
+			"resetLineNumbersWithPrefixedN": s.getBoolean(["feature", "resetLineNumbersWithPrefixedN"]),
+			"sdSupport": s.getBoolean(["feature", "sdSupport"])
 		},
 		"folder": {
 			"uploads": s.getBaseFolder("uploads"),
@@ -439,6 +487,9 @@ def setSettings():
 		if "feature" in data.keys():
 			if "gcodeViewer" in data["feature"].keys(): s.setBoolean(["feature", "gCodeVisualizer"], data["feature"]["gcodeViewer"])
 			if "waitForStart" in data["feature"].keys(): s.setBoolean(["feature", "waitForStartOnConnect"], data["feature"]["waitForStart"])
+			if "alwaysSendChecksum" in data["feature"].keys(): s.setBoolean(["feature", "alwaysSendChecksum"], data["feature"]["alwaysSendChecksum"])
+			if "resetLineNumbersWithPrefixedN" in data["feature"].keys(): s.setBoolean(["feature", "resetLineNumbersWithPrefixedN"], data["feature"]["resetLineNumbersWithPrefixedN"])
+			if "sdSupport" in data["feature"].keys(): s.setBoolean(["feature", "sdSupport"], data["feature"]["sdSupport"])
 
 		if "folder" in data.keys():
 			if "uploads" in data["folder"].keys(): s.setBaseFolder("uploads", data["folder"]["uploads"])
